requirements.txt
anthropic>=0.40.0
google-genai>=1.0.0
ollama>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
feedparser>=6.0.10
pytz>=2024.1
APScheduler>=3.10.4
